Refactor local semantic backend: introduce `local_runtime` for modular MetricFlow engine and SQL client handling. Streamline initialization and renderer logic.
